Fix merge on Synchronization
