--- internal/services/synchronizer.go
@@ -1,246 +1,0 @@
-package services
-
-import (
-	"strconv"
-	"time"
-
-	"github.com/NethermindEth/juno/internal/db/sync"
-	"github.com/NethermindEth/juno/pkg/feeder"
-	"github.com/NethermindEth/juno/pkg/felt"
-	"github.com/NethermindEth/juno/pkg/state"
-	"github.com/NethermindEth/juno/pkg/types"
-)
-
-type Synchronizer struct {
-	// stateManager is the state manager instance.
-	stateManager state.StateManager
-	// syncManager represents the sync manager
-	syncManager *sync.Manager
-	// feederClient is the feeder client instance.
-	feederClient *feeder.Client
-	// stateDiffCollector is the state diff collector instance.
-	stateDiffCollector StateDiffCollector
-}
-
-func NewSynchronizer(syncManager *sync.Manager, stateManager state.StateManager, feederClient *feeder.Client,
-	stateDiffCollector StateDiffCollector,
-) *Synchronizer {
-	return &Synchronizer{
-		stateManager:       stateManager,
-		syncManager:        syncManager,
-		feederClient:       feederClient,
-		stateDiffCollector: stateDiffCollector,
-	}
-}
-
-func (s *Synchronizer) UpdateBlock(blockNumber int64) (*feeder.StarknetBlock, error) {
-	block, err := s.feederClient.GetBlock("", strconv.FormatInt(blockNumber, 10))
-	if err != nil {
-		return nil, err
-	}
-
-	for _, txn := range block.Transactions {
-		err = s.updateTransactions(txn)
-		if err != nil {
-			iterations := 0
-			for {
-				time.Sleep(time.Second * 1)
-				if err = s.updateTransactions(txn); err == nil {
-					break
-				}
-				if iterations > 20 {
-					return nil, err
-				}
-				iterations++
-			}
-			if err != nil {
-				return nil, err
-			}
-		}
-	}
-	for i, txnReceipt := range block.TransactionReceipts {
-		err = s.updateTransactionReceipts(txnReceipt, block.Transactions[i].Type)
-		if err != nil {
-			iterations := 0
-			for {
-				time.Sleep(time.Second * 1)
-				if err = s.updateTransactionReceipts(txnReceipt, block.Transactions[i].Type); err == nil {
-					break
-				}
-				if iterations > 20 {
-					return nil, err
-				}
-				iterations++
-			}
-			if err != nil {
-				return nil, err
-			}
-		}
-	}
-
-	blockHash := new(felt.Felt).SetHex(block.BlockHash)
-	err = BlockService.StoreBlock(blockHash, feederBlockToDBBlock(block))
-	if err != nil {
-		return nil, err
-	}
-	return block, nil
-}
-
-func (s *Synchronizer) updateTransactions(txn feeder.TxnSpecificInfo) error {
-	transactionInfo, err := s.feederClient.GetTransaction(txn.TransactionHash, "")
-	if err != nil {
-		return err
-	}
-
-	transactionHash := new(felt.Felt).SetHex(transactionInfo.Transaction.TransactionHash)
-	// transactionHash := types.HexToTransactionHash(transactionInfo.Transaction.TransactionHash)
-	err = TransactionService.StoreTransaction(transactionHash, feederTransactionToDBTransaction(transactionInfo))
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (s *Synchronizer) updateTransactionReceipts(receipt feeder.TransactionExecution, txnType string) error {
-	txnReceipt, err := s.feederClient.GetTransactionReceipt(receipt.TransactionHash, "")
-	if err != nil {
-		return err
-	}
-	transactionHash := new(felt.Felt).SetHex(receipt.TransactionHash)
-	return TransactionService.StoreReceipt(transactionHash, feederTransactionToDBReceipt(txnReceipt, txnType))
-}
-
-func feederTransactionToDBReceipt(receipt *feeder.TransactionReceipt, txnType string) types.TxnReceipt {
-	common := types.TxnReceiptCommon{
-		TxnHash:     new(felt.Felt).SetHex(receipt.TransactionHash),
-		ActualFee:   new(felt.Felt).SetHex(receipt.ActualFee),
-		Status:      types.TxStatusValue[receipt.TxStatus],
-		StatusData:  receipt.TxStatus,
-		BlockHash:   new(felt.Felt).SetHex(receipt.BlockHash),
-		BlockNumber: uint64(receipt.BlockNumber),
-	}
-	switch txnType {
-	case "INVOKE_FUNCTION":
-		l2ToL1 := make([]*types.MsgToL1, 0)
-		for _, msg := range receipt.L2ToL1Messages {
-
-			payload := make([]*felt.Felt, 0)
-			for _, p := range msg.Payload {
-				payload = append(payload, new(felt.Felt).SetHex(p))
-			}
-
-			l2ToL1 = append(l2ToL1, &types.MsgToL1{
-				ToAddress:   types.HexToEthAddress(msg.ToAddress),
-				FromAddress: new(felt.Felt).SetHex(msg.FromAddress),
-				Payload:     payload,
-			})
-		}
-		payloadFromL1ToL2 := make([]*felt.Felt, 0)
-		for _, p := range receipt.L1ToL2Message.Payload {
-			payloadFromL1ToL2 = append(payloadFromL1ToL2, new(felt.Felt).SetHex(p))
-		}
-
-		events := make([]*types.Event, 0)
-
-		for _, event := range receipt.Events {
-
-			keys := make([]*felt.Felt, 0)
-			for _, p := range event.Keys {
-				keys = append(keys, new(felt.Felt).SetHex(p))
-			}
-			data := make([]*felt.Felt, 0)
-			for _, p := range event.Data {
-				data = append(data, new(felt.Felt).SetHex(p))
-			}
-
-			events = append(events, &types.Event{
-				FromAddress: new(felt.Felt).SetHex(event.FromAddress),
-				Keys:        keys,
-				Data:        data,
-			})
-		}
-
-		return &types.TxnInvokeReceipt{
-			TxnReceiptCommon: common,
-			MessagesSent:     l2ToL1,
-			L1OriginMessage: &types.MsgToL2{
-				FromAddress: types.HexToEthAddress(receipt.L1ToL2Message.FromAddress),
-				Payload:     payloadFromL1ToL2,
-				ToAddress:   new(felt.Felt).SetHex(receipt.L1ToL2Message.ToAddress),
-			},
-			Events: events,
-		}
-	case "DECLARE":
-		return &types.TxnDeclareReceipt{TxnReceiptCommon: common}
-	case "DEPLOY":
-		return &types.TxnInvokeReceipt{TxnReceiptCommon: common}
-
-	default:
-		return &common
-	}
-}
-
-// feederBlockToDBBlock convert the feeder block to the block stored in the database
-func feederBlockToDBBlock(b *feeder.StarknetBlock) *types.Block {
-	txnsHash := make([]*felt.Felt, 0)
-	for _, data := range b.Transactions {
-		txnsHash = append(txnsHash, new(felt.Felt).SetHex(data.TransactionHash))
-	}
-	status, _ := types.BlockStatusValue[b.Status]
-	return &types.Block{
-		BlockHash:   new(felt.Felt).SetHex(b.BlockHash),
-		BlockNumber: uint64(b.BlockNumber),
-		ParentHash:  new(felt.Felt).SetHex(b.ParentBlockHash),
-		Status:      status,
-		Sequencer:   new(felt.Felt).SetHex(b.SequencerAddress),
-		NewRoot:     new(felt.Felt).SetHex(b.StateRoot),
-		OldRoot:     new(felt.Felt).SetHex(b.OldStateRoot),
-		TimeStamp:   b.Timestamp,
-		TxCount:     uint64(len(b.Transactions)),
-		TxHashes:    txnsHash,
-	}
-}
-
-// feederTransactionToDBTransaction convert the feeder TransactionInfo to the transaction stored in DB
-func feederTransactionToDBTransaction(info *feeder.TransactionInfo) types.IsTransaction {
-	calldata := make([]*felt.Felt, 0)
-	for _, data := range info.Transaction.Calldata {
-		calldata = append(calldata, new(felt.Felt).SetHex(data))
-	}
-
-	switch info.Transaction.Type {
-	case "INVOKE_FUNCTION":
-		signature := make([]*felt.Felt, 0)
-		for _, data := range info.Transaction.Signature {
-			signature = append(signature, new(felt.Felt).SetHex(data))
-		}
-		return &types.TransactionInvoke{
-			Hash:               new(felt.Felt).SetHex(info.Transaction.TransactionHash),
-			ContractAddress:    new(felt.Felt).SetHex(info.Transaction.ContractAddress),
-			EntryPointSelector: new(felt.Felt).SetHex(info.Transaction.EntryPointSelector),
-			CallData:           calldata,
-			Signature:          signature,
-			MaxFee:             new(felt.Felt),
-		}
-	case "DECLARE":
-		signature := make([]*felt.Felt, 0)
-		for _, data := range info.Transaction.Signature {
-			signature = append(signature, new(felt.Felt).SetHex(data))
-		}
-		return &types.TransactionDeclare{
-			Hash:          new(felt.Felt).SetHex(info.Transaction.TransactionHash),
-			ClassHash:     new(felt.Felt).SetHex(info.Transaction.ContractAddress),
-			SenderAddress: new(felt.Felt).SetHex(info.Transaction.SenderAddress),
-			MaxFee:        new(felt.Felt).SetHex(info.Transaction.MaxFee),
-			Signature:     signature,
-			Nonce:         new(felt.Felt).SetHex(info.Transaction.Nonce),
-			Version:       new(felt.Felt).SetHex(info.Transaction.Version),
-		}
-	default:
-		return &types.TransactionDeploy{
-			Hash:                new(felt.Felt).SetHex(info.Transaction.TransactionHash),
-			ContractAddress:     new(felt.Felt).SetHex(info.Transaction.ContractAddress),
-			ConstructorCallData: calldata,
-		}
-	}
-}
